acvm::compiler
Module optimizers

Modules

common_subexpression 🔒

general 🔒

redundant_range 🔒

The redundant range constraint optimization pass aims to remove any BlackBoxFunc::Range opcodes which doesn’t result in additional restrictions on the values of witnesses.

unused_memory 🔒

Functions

optimize

Applies backend independent optimizations to a [Circuit].

optimize_internal 🔒

Applies backend independent optimizations to a [Circuit].