acvm::compiler::optimizers::common_subexpression
Module csat

Structs

CSatTransformer 🔒

A transformer which processes any Expressions to break them up such that they fit within the backend’s width.

Constants

MIN_EXPRESSION_WIDTH 🔒

Minimum width accepted by the CSatTransformer.

Functions

fits_in_one_identity 🔒

Checks if this expression can fit into one arithmetic identity