acvm::compiler
Struct AcirTransformationMap

pub struct AcirTransformationMap {
    old_indices_to_new_indices: HashMap<usize, Vec<usize>>,
}

This module can move and decompose acir opcodes into multiple opcodes. The transformation map allows consumers of this module to map metadata they had about the opcodes to the new opcode structure generated after the transformation. ACIR opcodes are stored inside a vector of opcodes. A transformation pass will generate a new vector of opcodes, but each opcode is the result of the transformation of an opcode in the original vector. So we simply keep track of the relation: index of the original opcode -> index of the new opcode in the new vector However we need a vector of new indexes for the map values in the case the old opcode is decomposed into multiple opcodes.

Fields
old_indices_to_new_indices: HashMap<usize, Vec<usize>>

Maps the old acir indices to the new acir indices

Implementations

impl AcirTransformationMap

fn new(acir_opcode_positions: &[usize]) -> Self

Builds a map from a vector of pointers to the old acir opcodes. The index in the vector is the new opcode index. The value of the vector is where the old opcode index was pointed. E.g: If acir_opcode_positions = 0,1,2,4,5,5,6 that means that old indices 0,1,2,4,5,5,6 are mapped to the new indexes: 0,1,2,3,4,5,6 This gives the following map: 0 -> 0 1 -> 1 2 -> 2 4 -> 3 5 -> [4, 5] 6 -> 6

pub fn new_locations( &self, old_location: OpcodeLocation, ) -> impl Iterator<Item = OpcodeLocation> + '_

Returns the new opcode location(s) corresponding to the old opcode. An OpcodeLocation contains the index of the opcode in the vector of opcodes This function returns the new OpcodeLocation by ‘updating’ the index within the given OpcodeLocation using the AcirTransformationMap. In fact, it does not update the given OpcodeLocation ‘in-memory’ but rather returns a new one, and even a vector of OpcodeLocation’s in case there are multiple new indexes corresponding to the old opcode index.

pub fn new_acir_locations( &self, old_location: AcirOpcodeLocation, ) -> impl Iterator<Item = AcirOpcodeLocation> + '_

This function is similar to new_locations(), but only deals with the AcirOpcodeLocation variant