acvm
Module pwg

Modules

arithmetic 🔒

blackbox 🔒

brillig 🔒

memory_op 🔒

Structs

ACVM

AcirCallWaitInfo

BrilligSolver

Specific solver for Brillig opcodes It maintains a Brillig VM that can execute the bytecode of the called brillig function

ForeignCallWaitInfo

Encapsulates a request from a Brillig VM process that encounters a foreign call opcode where the result of the foreign call has not yet been provided.

ProfilingSample

RawAssertionPayload

A dynamic assertion payload whose data has been resolved. This is instantiated upon hitting an assertion failure.

Enums

ACVMStatus

BrilligSolverStatus

ErrorLocation

Allows to point to a specific opcode as cause in errors. Some errors don’t have a specific opcode associated with them, or are created without one and added later.

OpcodeNotSolvable

OpcodeResolutionError

ResolvedAssertionPayload

Enumeration of possible resolved assertion payloads. This is instantiated upon hitting an assertion failure, and can either be static strings or dynamic payloads.

StepResult

Functions

any_witness_from_expression 🔒

get_value

Returns the concrete value for a particular expression If the value cannot be computed, it returns an ‘OpcodeNotSolvable’ error.

input_to_value

insert_value

Inserts value into the initial witness map under the index witness.

is_predicate_false 🔒

Returns true if the predicate is zero A predicate is used to indicate whether we should skip a certain operation. If we have a zero predicate it means the operation should be skipped.

witness_to_value

Type Aliases

ProfilingSamples