acvm::pwg::brillig
Function extract_failure_payload_from_memory

fn extract_failure_payload_from_memory<F: AcirField>(
    memory: &[MemoryValue<F>],
    revert_data_offset: usize,
    revert_data_size: usize,
) -> Option<ResolvedAssertionPayload<F>>

Extracts a ResolvedAssertionPayload from a block of memory of a Brillig VM instance.

Returns None if the amount of memory requested is zero.