acvm::pwg

Function is_predicate_false

pub(crate) fn is_predicate_false<F: AcirField>(
    witness: &WitnessMap<F>,
    predicate: &Expression<F>,
    opcode_location: &ErrorLocation,
) -> Result<bool, OpcodeResolutionError<F>>

Returns Ok(true) if the predicate is zero A predicate is used to indicate whether we should skip a certain operation. If we have a zero predicate it means the operation should be skipped.