acvm::pwg
Struct BrilligSolver

pub struct BrilligSolver<'b, F, B: BlackBoxFunctionSolver<F>> {
    vm: VM<'b, F, B>,
    acir_index: usize,
    pub function_id: BrilligFunctionId,
}

Specific solver for Brillig opcodes It maintains a Brillig VM that can execute the bytecode of the called brillig function

Fields
vm: VM<'b, F, B>acir_index: usizefunction_id: BrilligFunctionId

This id references which Brillig function within the main ACIR program we are solving. This is used for appropriately resolving errors as the ACIR program artifacts set up their Brillig debug metadata by function id.

Implementations

impl<'b, B: BlackBoxFunctionSolver<F>, F: AcirField> BrilligSolver<'b, F, B>

pub(super) fn zero_out_brillig_outputs( initial_witness: &mut WitnessMap<F>, outputs: &[BrilligOutputs], ) -> Result<(), OpcodeResolutionError<F>>

Assigns the zero value to all outputs of a given [brillig call][acir::circuit::opcodes::Opcode::BrilligCall].

pub(crate) fn new_call( initial_witness: &WitnessMap<F>, memory: &HashMap<BlockId, MemoryOpSolver<F>>, inputs: &'b [BrilligInputs<F>], brillig_bytecode: &'b [BrilligOpcode<F>], bb_solver: &'b B, acir_index: usize, brillig_function_id: BrilligFunctionId, profiling_active: bool, with_branch_to_feature_map: Option<&BranchToFeatureMap>, ) -> Result<Self, OpcodeResolutionError<F>>

Constructs a solver for a Brillig block given the bytecode and initial witness.

fn setup_brillig_vm( initial_witness: &WitnessMap<F>, memory: &HashMap<BlockId, MemoryOpSolver<F>>, inputs: &[BrilligInputs<F>], brillig_bytecode: &'b [BrilligOpcode<F>], bb_solver: &'b B, profiling_active: bool, with_branch_to_feature_map: Option<&BranchToFeatureMap>, ) -> Result<VM<'b, F, B>, OpcodeResolutionError<F>>

Get a BrilligVM for executing the provided bytecode

Reduce the input expressions into a known value, or error if they do not reduce to a value.

Instantiate the Brillig VM with the bytecode and the reduced inputs.

pub fn get_memory(&self) -> &[MemoryValue<F>]

pub fn write_memory_at(&mut self, ptr: usize, value: MemoryValue<F>)

pub fn get_call_stack(&self) -> Vec<usize>

pub fn get_fuzzing_trace(&self) -> Vec<u32>

pub(crate) fn solve( &mut self, ) -> Result<BrilligSolverStatus<F>, OpcodeResolutionError<F>>

pub fn step( &mut self, ) -> Result<BrilligSolverStatus<F>, OpcodeResolutionError<F>>

pub fn program_counter(&self) -> usize

fn handle_vm_status( &self, vm_status: VMStatus<F>, ) -> Result<BrilligSolverStatus<F>, OpcodeResolutionError<F>>

Returns the status of the Brillig VM as a ‘BrilligSolverStatus’ resolution. It may be finished, in-progress, failed, or may be waiting for results of a foreign call. Return the “resolution” to the caller who may choose to make subsequent calls (when it gets foreign call results for example).

pub(crate) fn finalize( self, witness: &mut WitnessMap<F>, outputs: &[BrilligOutputs], ) -> Result<(), OpcodeResolutionError<F>>

pub(crate) fn finalize_with_profiling( self, witness: &mut WitnessMap<F>, outputs: &[BrilligOutputs], ) -> Result<BrilligProfilingSamples, OpcodeResolutionError<F>>

Finalize the VM and return the profiling samples.

fn finalize_inner( &self, witness: &mut WitnessMap<F>, outputs: &[BrilligOutputs], ) -> Result<(), OpcodeResolutionError<F>>

Finalize the VM execution and write the outputs to the provided witness map.

fn write_brillig_outputs( &self, witness_map: &mut WitnessMap<F>, return_data_offset: usize, return_data_size: usize, outputs: &[BrilligOutputs], ) -> Result<(), OpcodeResolutionError<F>>

Write VM execution results into the witness map

pub fn resolve_pending_foreign_call( &mut self, foreign_call_result: ForeignCallResult<F>, )