acvm
Module compiler

Modules

optimizers 🔒

simulator 🔒

transformers 🔒

Structs

AcirTransformationMap

This module moves and decomposes acir opcodes. The transformation map allows consumers of this module to map metadata they had about the opcodes to the new opcode structure generated after the transformation.

CircuitSimulator

Simulate a symbolic solve for a circuit

Constants

MIN_EXPRESSION_WIDTH

Minimum width accepted by the CSatTransformer.

Functions

compile

Applies backend specific optimizations to a Circuit.

optimize

Applies backend independent optimizations to a Circuit.

transform

Applies backend specific optimizations to a Circuit.

transform_assert_messages 🔒

Update the assert messages to point to the new opcode locations.