acvm::compiler::optimizers
Module common_subexpression

Modules

csat 🔒

merge_expressions 🔒

Structs

WitnessFolder 🔒

Fold all witnesses in a circuit.

Constants

DEFAULT_MAX_TRANSFORMER_PASSES 🔒

We use multiple passes to stabilize the output in many cases

Functions

max_witness 🔒

Find the witness with the highest ID in the circuit.

transform_internal 🔒

Applies backend specific optimizations to a [Circuit].

transform_internal_once 🔒

Accepts an injected acir_opcode_positions to allow transformations to be applied directly after optimizations.